+++ clickhouse-20230522/alibabacloud_clickhouse20230522/models/_list_instance_linked_whitelist_templates_request.py
@@ -12,6 +12,8 @@ class ListInstanceLinkedWhitelistTemplatesRequest(DaraModel):
     ):
         # This parameter is required.
         self.dbinstance_id = dbinstance_id
+        # RegionId
+        # 
         # This parameter is required.
         self.region_id = region_id