+++ clickhouse-20230522/alibabacloud_clickhouse20230522/models/_list_whitelist_templates_request.py
@@ -14,6 +14,8 @@ class ListWhitelistTemplatesRequest(DaraModel):
     ):
         self.page_number = page_number
         self.page_size = page_size
+        # RegionId
+        # 
         # This parameter is required.
         self.region_id = region_id
         self.template_name = template_name