+++ clickhouse-20230522/alibabacloud_clickhouse20230522/models/_list_whitelist_templates_response_body.py
@@ -14,6 +14,7 @@ class ListWhitelistTemplatesResponseBody(DaraModel):
         request_id: str = None,
     ):
         self.data = data
+        # Id of the request
         self.request_id = request_id
 
     def validate(self):