+++ clickhouse-20230522/alibabacloud_clickhouse20230522/models/_update_whitelist_template_response_body.py
@@ -14,6 +14,7 @@ class UpdateWhitelistTemplateResponseBody(DaraModel):
         request_id: str = None,
     ):
         self.data = data
+        # Request ID。
         self.request_id = request_id
 
     def validate(self):